Remove extra labels types (#1314)
* Remove extra labels types

Remove the following labels types: INT32, UINT32, UINT64
and FLOAT32.

Fix all extra labels types occurrences in the project.


* Update CHANGELOG.md


* Delete unused helpers


* Convert passed values into remaining types


* Clarify func description

* Fix uint64 convertion


* Fix uint conversion


* Update OTLP exporter label types

+++ exporters/otlp/internal/otlptest/otlptest.go
@@ -57,7 +57,7 @@ func RunEndToEndTest(ctx context.Context, t *testing.T, exp *otlp.Exporter, mcTr
 	tp2 := sdktrace.NewTracerProvider(append(pOpts,
 		sdktrace.WithResource(resource.NewWithAttributes(
 			label.String("rk1", "rv12)"),
-			label.Float32("rk3", 6.5),
+			label.Float64("rk3", 6.5),
 		)))...)
 
 	tr1 := tp1.Tracer("test-tracer1")
